Dependancy on mouse clicks removed from main menu and sharing quest
Still to do, password and pause menu

+++ Assets/Scripts/Quests/SharingQuest/DoAClick.cs
@@ -19,7 +19,7 @@ public class DoAClick : MonoBehaviour
     {
         GameObject textBox = GameObject.FindGameObjectWithTag("TextBox");
         DialogueMessages d = textBox.GetComponent<DialogueMessages>();
-        accusation = (int)gameObject.name[3] - 48;
+        accusation = gameObject.name[3] - 48;
         transform.parent.gameObject.SetActive(false);
 
         d.switchButton(true);
